repo/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.16,<4.0
PyMySQL>=1.1,<2.0
cryptography>=42.0
argon2-cffi>=23.0
gunicorn>=22.0
openpyxl>=3.1,<4.0
croniter>=2.0,<4.0
pytest>=8.0
pytest-django>=4.8
